links, cleanup, template user without proxy

+++ etc/login.d/00_userinfo.sh
@@ -6,6 +6,14 @@
 
 if [ -n "$ad_connected" ]; then
 
+  if [ $USER == $template_user ]; then 
+    echo "Login as template user, no profile operations" >> /tmp/profile_exit
+    unset https_proxy
+    unset http_proxy
+    unset ftp_proxy
+    return
+  fi 
+
   userinfo="/tmp/$USER"
   [ -e "$userinfo" ] && rm -f "$userinfo"
